requirements.txt
asyncio==3.4.3
langchain==0.2.16
langchain-core==0.3.15
langchain-openai==0.2.6
openai==1.54.3
PyYAML==6.0.1
